a function that changes all lowercase letters of a string to uppercase

+++ 0x06-pointers_arrays_strings/5-string_toupper.c
@@ -10,7 +10,7 @@ char *string_toupper(char *str)
 
 	for (i = 0; str[i] != '\0'; i++)
 	{
-		str[i] = str[i];
+		str[i] = (int)str[i];
 	}
 	return (str);
 }
